Use idiomatic hcl inspection to identify target methods
