+++ internal/encryption/config/config_parse.go
@@ -36,6 +36,10 @@ func DecodeConfig(body hcl.Body, rng hcl.Range) (*EncryptionConfig, hcl.Diagnost
 				break
 			}
 		}
+
+		// Ensure that the key_provider address is valid
+		_, kpDiags := kp.Addr()
+		diags = diags.Extend(kpDiags)
 	}
 
 	for i, m := range cfg.MethodConfigs {
@@ -50,6 +54,10 @@ func DecodeConfig(body hcl.Body, rng hcl.Range) (*EncryptionConfig, hcl.Diagnost
 				break
 			}
 		}
+
+		// Ensure that the method address is valid
+		_, mDiags := m.Addr()
+		diags = diags.Extend(mDiags)
 	}
 
 	if cfg.Remote != nil {
